An den Beans wegen der ID rumgeschraubt, weil die INTEGER waren anstatt LONG.... Bei der Datenbank weitere Methoden hinzugefügt.
git-svn-id: https://randi2.svn.sourceforge.net/svnroot/randi2/randi/trunk/RANDI2@911 ccfe945e-4522-0410-b90e-900a8cb55c58

+++ src/de/randi2/model/fachklassen/Studienarm.java
@@ -32,7 +32,7 @@ public final class Studienarm {
 	 * @throws DatenbankFehlerException
 	 *             falls Fehler bei dem Vorgang auftraten.
 	 */
-	public static StudienarmBean get(int id) throws DatenbankFehlerException {
+	public static StudienarmBean get(long id) throws DatenbankFehlerException {
 		StudienarmBean nullBean = new StudienarmBean();
 		nullBean.setFilter(true);
 		return DatenbankFactory.getAktuelleDBInstanz().suchenObjektID(id,
